feat: Add comment service

+++ forum_back/src/main/kotlin/com/example/forum_back/entity/Comment.kt
@@ -10,7 +10,7 @@ import jakarta.persistence.Id
 class Comment (
     var description: String,
     var author: String,
-    val forumId: Int,
+    val forumId: Long,
     @Id
     @GeneratedValue(strategy = IDENTITY)
     val id: Long? = null
